'Codeless' UI test generation - Add: Makefile commands to use 'playright codegen' - Add: New test file for N3C recommended (WIP) - Add: frontend/tests/utils.js (refactored func into here from another file)
